chess quant v2/chessquant-ml-labs-v2/chessquant-ml-labs-v2/pyproject.toml
[tool.poetry]
name = "chessquant-ml-labs-v2"
version = "0.1.0"
description = "Local ML lab for retraining the ChessQuant tilt detector"
authors = ["John Scott <chatgpt4.mm.acc.2@gmail.com>"]
packages = [{ include = "chessquant_ml", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
polars = "^1.9.0"
duckdb = "^1.1.0"
python-chess = "^1.999"
zstandard = "^0.23.0"
pyarrow = "^18.0.0"
numpy = "^2.1.0"
pandas = "^2.2.0"
scikit-learn = "^1.5.0"
xgboost = "^2.1.0"
onnx = "^1.17.0"
skl2onnx = "^1.17.0"
onnxruntime = "^1.19.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.0"
mlflow = "^2.17.0"
wandb = "^0.18.0"
typer = "0.15.3"
joblib = "^1.4.2"
tqdm = "^4.67.3"
click = "8.1.7"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
pytest = "^8.3.0"

[tool.poetry.scripts]
chessquant-ml = "chessquant_ml.cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"